requirements.txt
tensorflow==2.3.0
streamlit==0.64.0
numpy==1.18.5
Pillow==7.2.0
